# daebbang-api/src/main/resources/static/swagger-ui/openapi3.yaml
openapi: 3.0.1
info:
  title: 대빵 언니 API 서버 API 명세서
  description: 대빵 언니 API 서버 명세서입니다.
  version: 1.0.0
servers:
- url: https://api.daebbang-sister.shop
  description: 운영 서버
- url: http://localhost:8080
  description: 로컬 개발 서버
...
paths:
  /v1/carts:
    get:
      tags:
      - Cart
      summary: 장바구니 조회
      description: 로그인한 회원의 장바구니를 커서 기반 페이지네이션으로 조회합니다. cursor 생략 시 최신 항목부터 반환합니다.
      operationId: cart/get-carts-
      parameters:
      - name: cursor
        in: query
        description: 마지막으로 조회한 cartId (첫 페이지 요청 시 생략)
        required: false
        schema:
          type: string
      - name: size
        in: query
        description: "페이지 크기 (기본값: 8)"
        required: false
        schema:
          type: string
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartPageResponse'
              examples:
                cart/get-carts-first-page:
                  value: "{\"success\":true,\"status\":200,\"message\":\"조회에 성공하였습\
                    니다.\",\"data\":{\"carts\":[{\"cartId\":10,\"productId\":100,\"\
                    quantity\":1,\"productName\":\"테스트 상품\",\"originalPrice\":30000,\"\
                    discountPrice\":30000,\"discountRate\":null,\"color\":\"블랙\",\"\
                    size\":\"M\",\"mainImageUrl\":\"https://example.com/image.jpg\"\
                    },{\"cartId\":9,\"productId\":100,\"quantity\":1,\"productName\"\
                    :\"테스트 상품\",\"originalPrice\":30000,\"discountPrice\":30000,\"\
                    discountRate\":null,\"color\":\"블랙\",\"size\":\"M\",\"mainImageUrl\"\
                    :\"https://example.com/image.jpg\"},{\"cartId\":8,\"productId\"\
                    :100,\"quantity\":1,\"productName\":\"테스트 상품\",\"originalPrice\"\
                    :30000,\"discountPrice\":30000,\"discountRate\":null,\"color\"\
                    :\"블랙\",\"size\":\"M\",\"mainImageUrl\":\"https://example.com/image.jpg\"\
                    },{\"cartId\":7,\"productId\":100,\"quantity\":1,\"productName\"\
                    :\"테스트 상품\",\"originalPrice\":30000,\"discountPrice\":30000,\"\
                    discountRate\":null,\"color\":\"블랙\",\"size\":\"M\",\"mainImageUrl\"\
                    :\"https://example.com/image.jpg\"},{\"cartId\":6,\"productId\"\
                    :100,\"quantity\":1,\"productName\":\"테스트 상품\",\"originalPrice\"\
                    :30000,\"discountPrice\":30000,\"discountRate\":null,\"color\"\
                    :\"블랙\",\"size\":\"M\",\"mainImageUrl\":\"https://example.com/image.jpg\"\
                    },{\"cartId\":5,\"productId\":100,\"quantity\":1,\"productName\"\
                    :\"테스트 상품\",\"originalPrice\":30000,\"discountPrice\":30000,\"\
                    discountRate\":null,\"color\":\"블랙\",\"size\":\"M\",\"mainImageUrl\"\
                    :\"https://example.com/image.jpg\"},{\"cartId\":4,\"productId\"\
                    :100,\"quantity\":1,\"productName\":\"테스트 상품\",\"originalPrice\"\
                    :30000,\"discountPrice\":30000,\"discountRate\":null,\"color\"\
                    :\"블랙\",\"size\":\"M\",\"mainImageUrl\":\"https://example.com/image.jpg\"\
                    },{\"cartId\":3,\"productId\":100,\"quantity\":1,\"productName\"\
                    :\"테스트 상품\",\"originalPrice\":30000,\"discountPrice\":30000,\"\
                    discountRate\":null,\"color\":\"블랙\",\"size\":\"M\",\"mainImageUrl\"\
                    :\"https://example.com/image.jpg\"}],\"nextCursor\":3,\"hasNext\"\
                    :true}}"
                cart/get-carts-last-page:
                  value: "{\"success\":true,\"status\":200,\"message\":\"조회에 성공하였습\
                    니다.\",\"data\":{\"carts\":[{\"cartId\":3,\"productId\":100,\"\
                    quantity\":1,\"productName\":\"테스트 상품\",\"originalPrice\":30000,\"\
                    discountPrice\":30000,\"discountRate\":null,\"color\":\"블랙\",\"\
                    size\":\"M\",\"mainImageUrl\":\"https://example.com/image.jpg\"\
                    },{\"cartId\":2,\"productId\":100,\"quantity\":1,\"productName\"\
                    :\"테스트 상품\",\"originalPrice\":30000,\"discountPrice\":30000,\"\
                    discountRate\":null,\"color\":\"블랙\",\"size\":\"M\",\"mainImageUrl\"\
                    :\"https://example.com/image.jpg\"},{\"cartId\":1,\"productId\"\
                    :100,\"quantity\":1,\"productName\":\"테스트 상품\",\"originalPrice\"\
                    :30000,\"discountPrice\":30000,\"discountRate\":null,\"color\"\
                    :\"블랙\",\"size\":\"M\",\"mainImageUrl\":\"https://example.com/image.jpg\"\
                    }],\"nextCursor\":null,\"hasNext\":false}}"
    post:
      tags:
      - Cart
      summary: 장바구니 추가
      description: 장바구니에 상품을 추가합니다. 이미 담긴 상품이면 수량을 합산합니다. 단건 또는 다건 모두 리스트 형태로 전달합니다.
      operationId: cart/save-carts
      requestBody:
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: '#/components/schemas/CartSaveRequest'
            examples:
              cart/save-carts:
                value: "[{\"productDetailId\":1,\"quantity\":2}]"
              cart/save-carts-invalid-quantity:
                value: "[{\"productDetailId\":1,\"quantity\":0}]"
              cart/save-carts-multiple:
                value: "[{\"productDetailId\":1,\"quantity\":2},{\"productDetailId\"\
                  :2,\"quantity\":1},{\"productDetailId\":3,\"quantity\":3}]"
      responses:
        "201":
          description: "201"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              examples:
                cart/save-carts:
                  value: "{\"success\":true,\"status\":201,\"message\":\"장바구니 추가에\
                    \ 성공하였습니다.\",\"data\":null}"
                cart/save-carts-multiple:
                  value: "{\"success\":true,\"status\":201,\"message\":\"장바구니 추가에\
                    \ 성공하였습니다.\",\"data\":null}"
        "400":
          description: "400"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                cart/save-carts-invalid-quantity:
                  value: "{\"success\":false,\"status\":400,\"message\":\"잘못된 입력값입\
                    니다.\",\"data\":null,\"errors\":[{\"field\":\"quantity\",\"message\"\
                    :\"수량은 최소 한 개 이상 선택해야 합니다.\"}]}"
    delete:
      tags:
      - Cart
      summary: 장바구니 선택 삭제
      description: 선택한 장바구니 항목들을 삭제합니다. 본인 소유 항목만 삭제됩니다.
      operationId: cart/delete-carts
      parameters:
      - name: ids
        in: query
        description: 삭제할 장바구니 항목 ID 목록 (1개 이상)
        required: true
        schema:
          type: string
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              examples:
                cart/delete-carts:
                  value: "{\"success\":true,\"status\":200,\"message\":\"장바구니 삭제에\
                    \ 성공하였습니다.\",\"data\":null}"
...
  /v1/wish-lists:
    get:
      tags:
      - WishList
      summary: 위시리스트 조회
      description: 로그인한 회원의 위시리스트를 페이지네이션으로 조회합니다. 최신 추가 순으로 정렬됩니다.
      operationId: wish-lists/list-0
      parameters:
      - name: page
        in: query
        description: "페이지 번호 (0부터 시작, 기본값: 0)"
        required: false
        schema:
          type: string
      - name: size
        in: query
        description: "페이지 크기 (기본값: 20)"
        required: false
        schema:
          type: string
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WishListPageResponse'
              examples:
                wish-lists/list-01-success:
                  value: "{\"success\":true,\"status\":200,\"message\":\"조회에 성공하였습\
                    니다.\",\"data\":{\"content\":[{\"wishListId\":2,\"productId\":100,\"\
                    mainImageUrl\":\"https://example.com/image.jpg\",\"productName\"\
                    :\"테스트 상품\",\"originalPrice\":30000,\"discountRate\":null,\"sellingPrice\"\
                    :30000},{\"wishListId\":1,\"productId\":200,\"mainImageUrl\":\"\
                    https://example.com/sale-image.jpg\",\"productName\":\"할인 상품\"\
                    ,\"originalPrice\":50000,\"discountRate\":10,\"sellingPrice\"\
                    :45000}],\"pageNumber\":0,\"pageSize\":10,\"totalElements\":2,\"\
                    totalPages\":1,\"last\":true}}"
                wish-lists/list-02-empty:
                  value: "{\"success\":true,\"status\":200,\"message\":\"조회에 성공하였습\
                    니다.\",\"data\":{\"content\":[],\"pageNumber\":0,\"pageSize\":10,\"\
                    totalElements\":0,\"totalPages\":0,\"last\":true}}"
    post:
      tags:
      - WishList
      summary: 위시리스트 추가
      description: 상품 상세 페이지에서 상품을 위시리스트에 추가합니다. 이미 추가된 상품은 409를 반환합니다.
      operationId: wish-lists/add-0
      requestBody:
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: '#/components/schemas/WishListSaveRequest'
            examples:
              wish-lists/add-01-success:
                value: "{\"productId\":100}"
              wish-lists/add-02-already-exists:
                value: "{\"productId\":100}"
      responses:
        "201":
          description: "201"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              examples:
                wish-lists/add-01-success:
                  value: "{\"success\":true,\"status\":201,\"message\":\"위시리스트 추가에\
                    \ 성공하였습니다.\",\"data\":null}"
        "409":
          description: "409"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                wish-lists/add-02-already-exists:
                  value: "{\"success\":false,\"status\":409,\"message\":\"이미 위시리스트\
                    에 추가된 상품입니다.\",\"data\":null}"
    delete:
      tags:
      - WishList
      summary: 위시리스트 선택 삭제
      description: "선택한 위시리스트 항목들을 삭제합니다. 본인 소유 항목만 삭제되며, 단건 삭제도 동일하게 처리합니다."
      operationId: wish-lists/delete-01-selected
      parameters:
      - name: ids
        in: query
        description: 삭제할 위시리스트 항목 ID 목록 (1개 이상)
        required: true
        schema:
          type: string
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              examples:
                wish-lists/delete-01-selected:
                  value: "{\"success\":true,\"status\":200,\"message\":\"위시리스트 삭제에\
                    \ 성공하였습니다.\",\"data\":null}"
...
  /v1/carts/{cartId}:
    patch:
      tags:
      - Cart
      summary: 장바구니 수정
      description: 장바구니 항목의 색상/사이즈(productDetailId)와 수량을 수정합니다.
      operationId: cart/update-cart
      parameters:
      - name: cartId
        in: path
        description: 수정할 장바구니 항목 ID
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: '#/components/schemas/CartUpdateRequest'
            examples:
              cart/update-cart:
                value: "{\"productDetailsId\":2,\"quantity\":3}"
              cart/update-cart-not-found:
                value: "{\"productDetailsId\":2,\"quantity\":3}"
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              examples:
                cart/update-cart:
                  value: "{\"success\":true,\"status\":200,\"message\":\"장바구니 수정에\
                    \ 성공하였습니다.\",\"data\":null}"
        "404":
          description: "404"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                cart/update-cart-not-found:
                  value: "{\"success\":false,\"status\":404,\"message\":\"존재하지 않는\
                    \ 카트 번호입니다.\",\"data\":null}"
...
components:
  schemas:
    ErrorResponse:
      title: ErrorResponse
      type: object
      properties:
        data:
          type: object
          description: 응답 데이터 (없음)
        success:
          type: boolean
          description: 성공 여부 (false)
        message:
          type: string
          description: 오류 메시지
        errors:
          type: array
          description: 필드 유효성 검증 오류 목록
          items:
            type: object
            properties:
              field:
                type: string
                description: 오류 필드명
              message:
                type: string
                description: 오류 메시지
        status:
          type: number
          description: HTTP 상태 코드 (400)
...
    CartPageResponse:
      title: CartPageResponse
      type: object
      properties:
        data:
          type: object
          properties:
            nextCursor:
              type: object
              description: 다음 페이지 cursor (null)
            carts:
              type: array
              description: 장바구니 항목 목록
              items:
                type: object
                properties:
                  discountRate:
                    type: object
                    description: 할인율 (%)
                  quantity:
                    type: number
                    description: 수량
                  size:
                    type: string
                    description: 사이즈
                  color:
                    type: string
                    description: 색상
                  originalPrice:
                    type: number
                    description: 정가
                  productId:
                    type: number
                    description: 상품 ID
                  cartId:
                    type: number
                    description: 장바구니 항목 ID
                  discountPrice:
                    type: number
                    description: 할인 적용가
                  mainImageUrl:
                    type: string
                    description: 대표 이미지 URL
                  productName:
                    type: string
                    description: 상품명
            hasNext:
              type: boolean
              description: 다음 페이지 존재 여부
        success:
          type: boolean
          description: 성공 여부
        message:
          type: string
          description: 응답 메시지
        status:
          type: number
          description: HTTP 상태 코드
...
    CartUpdateRequest:
      title: CartUpdateRequest
      type: object
      properties:
        productDetailsId:
          type: number
          description: 변경할 상품 상세 ID
        quantity:
          type: number
          description: 변경할 수량
    WishListPageResponse:
      title: WishListPageResponse
      type: object
      properties:
        data:
          type: object
          properties:
            pageNumber:
              type: number
              description: 현재 페이지 번호
            last:
              type: boolean
              description: 마지막 페이지 여부
            totalPages:
              type: number
              description: 전체 페이지 수
            pageSize:
              type: number
              description: 페이지 크기
            content:
              type: array
              description: 위시리스트 항목 목록 (빈 배열)
              items:
                type: object
                properties:
                  discountRate:
                    type: object
                    description: 할인율 (%) - 할인 없으면 null
                  sellingPrice:
                    type: number
                    description: 판매가 (할인 없으면 정가와 동일)
                  wishListId:
                    type: number
                    description: 위시리스트 항목 ID (삭제 시 사용)
                  originalPrice:
                    type: number
                    description: 정가
                  productId:
                    type: number
                    description: 상품 ID
                  mainImageUrl:
                    type: string
                    description: 대표 이미지 URL
                  productName:
                    type: string
                    description: 상품명
            totalElements:
              type: number
              description: 전체 항목 수
        success:
          type: boolean
          description: 성공 여부
        message:
          type: string
          description: 응답 메시지
        status:
          type: number
          description: HTTP 상태 코드
...
    WishListSaveRequest:
      title: WishListSaveRequest
      type: object
      properties:
        productId:
          type: number
          description: 추가할 상품 ID
...
    SuccessResponse:
      title: SuccessResponse
      type: object
      properties:
        data:
          type: object
          description: 응답 데이터 (없음)
        success:
          type: boolean
          description: 성공 여부
        message:
          type: string
          description: 응답 메시지
        status:
          type: number
          description: HTTP 상태 코드
...
    CartSaveRequest:
      title: CartSaveRequest
      type: array
      items:
        type: object
        properties:
          quantity:
            type: number
            description: 수량 (최소 1)
          productDetailId:
            type: number
            description: 상품 상세 ID
...